add additional db methods and set up test file

+++ business/db_interfaces.py
@@ -2,7 +2,8 @@ import mysql.connector
 from mysql.connector.errors import *
 from mysql.connector import connect
 
-# import Object Classes from folder to allow for testing
+# TODO import Object Classes from folder to allow for testing
+# TODO finalise DB populator method
 
 import sys, os
 sys.path.append('./') #allows for usage of 'admin.py' contents
@@ -45,13 +46,13 @@ class MySQLWorkbenchInterface():
             raise Exception("Error occured while searching for database!")
 
 class DatabaseInterface() :
-    def __init__(self) :
+    def __init__(self, db_name) :
         # intitialise connection to database and set up cursor to execute SQL queries
-        self.connection = connect(host = "localhost", user = DBUSERNAME, password = DBPASSWORD, database = "alpha_db")
+        self.connection = connect(host = "localhost", user = DBUSERNAME, password = DBPASSWORD, database = f"{db_name}")
         self.cursor = self.connection.cursor(buffered=True)
 
         # create the 'students' table, if already exist then command is ignored
-        try: self.cursor.execute("CREATE TABLE users (fullname VARCHAR(50), studentID VARCHAR(36) ,email VARCHAR(60), password VARCHAR(30))")
+        try: self.cursor.execute("CREATE TABLE students (fullname VARCHAR(50), studentID VARCHAR(36) ,email VARCHAR(60), password VARCHAR(30))")
         except ProgrammingError: pass
 
         # create the 'employers' table, if already exist then command is ignored
@@ -60,6 +61,7 @@ class DatabaseInterface() :
 
     # create new entry into the student table
     def add_student(self, student : Student):
+        if self.students_exists(student) : raise Exception("Student with given credentials already exists")
         try:
             query = f"INSERT INTO students (fullname, studentID, email, username, password) VALUES ('{student.get_fullname()}', '{student.get_id()}', '{student.get_email()}', '{student.get_password()}')"
             self.cursor.execute(query)
@@ -70,6 +72,7 @@ class DatabaseInterface() :
 
     # create new entry into the employers table
     def add_employer(self, employer : Employer):
+        if self.employer_exists(employer) : raise Exception("Employer with given credentials already exists")
         try:
             query = f"INSERT INTO employers (name, employerID, email) VALUES ('{employer.get_company_name()}', '{employer.get_id()}', '{employer.get_email()}')"
             self.cursor.execute(query)
@@ -80,16 +83,75 @@ class DatabaseInterface() :
 
     # delete entry from the student table
     def delete_student(self, student : Student) :
-        query = f"DELETE FROM students WHERE stduentID='{student.get_id()}'"
-        self.cursor.execute(query)
-        self.db.commit()
+        if not self.students_exists(student) : raise Exception("Student does not exist")
+        try:
+            query = f"DELETE FROM students WHERE stduentID='{student.get_id()}'"
+            self.cursor.execute(query)
+            self.db.commit()
+        except ProgrammingError as error:
+            raise Exception("Failed to delete student from database")
+            print(error)
 
     # delete entry from the employers table
     def delete_employer(self, employer : Employer):
-        query = f"DELETE FROM employers WHERE employerID='{employer.get_id()}'"
-        self.cursor.execute(query)
-        self.db.commit()
+        if not self.employer_exists(employer) : raise Exception("Employer does not exist")
+        try:
+            query = f"DELETE FROM employers WHERE employerID='{employer.get_id()}'"
+            self.cursor.execute(query)
+            self.db.commit()
+        except ProgrammingError as error:
+            raise Exception("Failed to delete employer from database")
+            print(error)
 
-#
-def populate_db() :
-    pass
+    # returns True if the student is already in the table, False otherwise
+    def student_exists(self, student : Student):
+        try:
+            query = f"SELECT * FROM students WHERE id='{student.get_id()}'"
+            self.cursor.execute(query)
+            rows = self.cursor.fetchall()
+            return len(rows) == 1
+        except ProgrammingError as error:
+            raise Exception("Error while checking for student presence")
+
+    # returns True if the employer is already in the table, False otherwise
+    def employer_exists(self, employer : Employer):
+        try:
+            query = f"SELECT * FROM employers WHERE id='{employer.get_id()}'"
+            self.cursor.execute(query)
+            rows = self.cursor.fetchall()
+            return len(rows) == 1
+        except ProgrammingError as error:
+            raise Exception("Error while checking for employer presence")
+
+    def show_table_rows(self, table_name):
+        # prints the contents of the given table
+        try:
+            self.cursor.execute(f"SELECT * FROM {table_name}")
+            return [ row for row in list(self.cursor) ]
+        except ProgrammingError as error:
+            print("Error while trying to display table data. Make sure table exists")
+
+
+# populate database given that user calling method is authorised (only the dev team)
+# returns True if method was successfully executed else False
+def populate_db(db_name) :
+    dev_password = str(input("Enter development password : "))
+    if (dev_password != "AlphaTeamBestTeam!"):
+        raise Exception("Missing credentials")
+        return False
+
+    # set up interfaces and reset working environment
+    workbench = MySQLWorkbenchInterface()
+    database = DatabaseInterface("alpha_db")
+    workbench.destroy_db("alpha_db") 
+    workbench.create_db("alpha_db")
+    
+    # populate students table with 50 dummy entries
+    for i in range(0, 50):
+        pass
+
+    #populate employers table with 15 dummy entries
+    for i in range(0, 15):
+        pass
+
+    # NOTE : populate process will have a 'for' loop for each entity in the final DB schema
